feat: support detection of node/npm apps
This adds some more logic to also detect node-based frontend apps. The
web-servers buildpack already takes care of running `npm build` but in
case that happens we need to serve the resulting artifacts from
`/workspace/build`. We do that by leveraging the libnodejs package to
detect if we have a nodejs app and then we set the build directory as the
webroot.