fix: disable absolute redirects
nginx will automatically redirect to paths with a trailing slash it
is missing. But this lead to issues where nginx would redirect from
HTTPS back to HTTP if it's running behind a proxy which already
terminates TLS. As a workaround we configure nginx to do relative
redirects that don't contain the scheme.